requirements.txt
matplotlib==3.3.2
pyyaml==5.3.1
Telethon==1.16.4
# spacy==2.3.2
natasha==1.3.0